refactor: group review components in 1 dir
