+++ src/components/flashcard/main/swipe-action.tsx
